requirements.txt
fastapi==0.135.1
uvicorn==0.42.0

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.1
scipy==1.15.3

xgboost==3.2.0
joblib==1.5.3

pydantic==2.12.3
python-multipart==0.0.22

python-dotenv

pika
sqlalchemy
psycopg2-binary
email-validator